LRIS-22555 Use correct column type

+++ modules/private/apps/osb-ldn/osb-ldn-generator/osb-ldn-generator/src/main/java/com/liferay/osb/ldn/generator/layout/LayoutVersion.java
@@ -98,7 +98,7 @@ public class LayoutVersion {
 		if (expandoColumn == null) {
 			_expandoColumnLocalService.addColumn(
 				expandoTable.getTableId(), _COLUMN_NAME,
-				ExpandoColumnConstants.LONG);
+				ExpandoColumnConstants.INTEGER);
 		}
 	}
